+++ User_Files/1_Middleware/Algorithm/PID/alg_pid.h
@@ -121,8 +121,8 @@ typedef struct _PID_TypeDef
 
 void PID_Init(
     PID_TypeDef *pid,
-    uint16_t max_out,
-    uint16_t intergral_limit,
+    float max_out,
+    float intergral_limit,
     float deadband,
 
     float kp,
